+++ ecs_service/ecs_service.tf
@@ -12,7 +12,7 @@ data "aws_iam_role" "main" {
 }
 
 data "aws_lb_target_group" "ecs" {
-  name = "${local.name}-ecs"
+  name = var.load_balancer_target_group_name == null ? "${local.name}-ecs" : var.load_balancer_target_group_name
 }
 
 data "aws_vpc" "main" {
@@ -43,7 +43,7 @@ resource "aws_ecs_service" "main" {
   name            = var.container_name
   cluster         = data.aws_ecs_cluster.main.id
   task_definition = aws_ecs_task_definition.main.arn
-  desired_count   = 1
+  desired_count   = var.desired_count
   launch_type     = "FARGATE"
 
   network_configuration {
@@ -57,11 +57,11 @@ resource "aws_ecs_service" "main" {
 
   # not required if you don't want to use a load balancer, e.g. for Sidekiq
   dynamic "load_balancer" {
-    for_each = var.app_container_port == null ? [] : [{
+    for_each = var.with_load_balancer ? [{
       target_group_arn = data.aws_lb_target_group.ecs.arn,
       container_name   = var.container_name,
       container_port   = var.app_container_port
-    }]
+    }] : []
     content {
       target_group_arn = load_balancer.value.target_group_arn
       container_name   = load_balancer.value.container_name